X509 classes refactor
